rename bootstrapper -> enroller
