--- apiserver/bootstrapper/bootstrapper.go
@@ -1,7 +1,0 @@
-package bootstrapper
-
-import "net/http"
-
-type bootstrapper struct {
-	Client *http.Client
-}
