+++ apiserver/enroller/enroller.go
@@ -1,0 +1,7 @@
+package enroller
+
+import "net/http"
+
+type enroller struct {
+	Client *http.Client
+}
